355 mergecubes uses default geotrellis partitioner in certain cases which causes oom errors (#356)
* add ConfigurableSpatialPartitionerReduceZ for more granular partition sizes #355

* replace default partitioner with getPartitionerIndexForMaxPartitionSize in merge_cubes #355

The default SpacePartitioner[K](kb) uses 16x16 tiles which uses too much memory if there are many bands, dates or if the tiles are large. This commit introduces a partitioner that attempts to control the partition size based on a maximum memory footprint.

* fix tilesize default in merge cubes #355

* add documentation to getPartitionerIndexForMaxPartitionSize #355

* revert manyBands case in outerJoin #355

* use worst case as default for number of bands #355

With this default, for a maximum partition size of 500 Mb the following would be true:
* 20 bands - 500Mb partitions
* 40 bands - 1Gb partitions
* 10 bands - 250Mb partitions

This should be a good balance, since 40 bands is quite extreme.