use sparseStitch in stitchAndWriteToTiff #413
The tiles we pass to stitch have all empty tiles filtered out. Stitch() just stitched them together without adding the empty tiles in between.

We assume that there are always valid tiles and that the layout.extent is correct. But in the case that sparseStitch fails we try to recover by writing an empty tile anyway. This is however an actual error somewhere in the execution flow so I do log it as an error for the user.